bug corregidos y ladingpage

+++ frontend/src/pages/SeleccionEstudiante1.js
@@ -58,10 +58,14 @@ const SeleccionEstudiante1 = () => {
     const [continuetotal, setContinueTotal] = useState(0);
 
     const [selectedRadio, setSelectedRadio] = useState(null);
+    const [recordname, setRecordName] = useState(null);
     
       const handleCheckboxChange = (event) => {
-        setSelectedId(parseInt(event.target.value));
-        setSelectedRadio(parseInt(event.target.value));
+        const value = event.target.value;
+        const [recordid, subjectname] = value.split('_');
+        setRecordName(subjectname);
+        setSelectedId(parseInt(recordid));
+        setSelectedRadio(parseInt(recordid));
       };    
     
     function handleChange(e) {
@@ -91,9 +95,12 @@ const SeleccionEstudiante1 = () => {
         window.alert('Debes seleccionar una asignatura');
         return;
       }
+        // Show confirmation messagebox
+  const confirmed = window.confirm(`¿Está seguro de que desea retirar la materia ${recordname}?`);
+
+  if (confirmed) {
       setSelectedId(null);
       setSelectedRadio(null);
-      window.alert(selectedId);
       axios.delete(`http://localhost:8000/api/v1/subject-cycles/${selectedId}/retire/`, config)
       
         .then(response => {
@@ -105,7 +112,7 @@ const SeleccionEstudiante1 = () => {
           window.alert ('Ocurrio un error al retirar la materia');
           console.error('Ocurrió un error al eliminar el recurso:', error);
         })
-        
+      }
     }
 
   
@@ -228,12 +235,6 @@ const SeleccionEstudiante1 = () => {
             </div>
             <button className={styles.rectangleParent}>
               <div className={styles.groupChild} />
-              <img
-                className={styles.vuesaxlinearnotificationIcon}
-                alt=""
-                src="/vuesaxlinearnotification1.svg"
-              />
-              <img className={styles.groupItem} alt="" src="/ellipse-48.svg" />
             </button>
           </div>
         </div>
@@ -281,7 +282,7 @@ const SeleccionEstudiante1 = () => {
                   {subjectcode1} - {subjectname1}
                 </div>
               </div> 
-              <input className={styles.boxunchecked} type="radio"   name="radio" value={recordid1} checked={selectedRadio === recordid1} onChange={handleCheckboxChange} />
+              <input className={styles.boxunchecked} type="radio"   name="radio" value={`${recordid1}_${subjectname1}`} checked={selectedRadio === recordid1} onChange={handleCheckboxChange} />
             </div>)}
             {(jValue+1 < datalength) && (
             <div className={styles.div3}>
@@ -291,7 +292,7 @@ const SeleccionEstudiante1 = () => {
               </div>
               <div className={styles.div4}>{subjectcredit2}</div>
               <div className={styles.div5}>{subjectlab2? 'Si':'No'}</div>
-              <input className={styles.boxunchecked1} type="radio"   name="radio" value={recordid2} checked={selectedRadio === recordid2} onChange={handleCheckboxChange} />
+              <input className={styles.boxunchecked1} type="radio"   name="radio" value={`${recordid2}_${subjectname2}`} checked={selectedRadio === recordid2} onChange={handleCheckboxChange} />
             </div>)}
             {(jValue+2 < datalength) && (
             <div className={styles.div6}>
@@ -303,7 +304,7 @@ const SeleccionEstudiante1 = () => {
                 <div className={styles.div7}>{subjectcredit3}</div>
                 <div className={styles.div8}>{subjectlab3? 'Si':'No'}</div>
               </div>
-              <input className={styles.boxunchecked4} type="radio"   name="radio" value={recordid3} checked={selectedRadio === recordid3} onChange={handleCheckboxChange}/>
+              <input className={styles.boxunchecked4} type="radio"   name="radio" value={`${recordid3}_${subjectname3}`} checked={selectedRadio === recordid3} onChange={handleCheckboxChange}/>
             </div>)}
             {(jValue+3 < datalength) && (
             <div className={styles.div9}>
@@ -315,7 +316,7 @@ const SeleccionEstudiante1 = () => {
               </div>
               <div className={styles.div10}>{subjectlab4? 'Si':'No'}</div>
               <div className={styles.div11}>{subjectcredit4}</div>
-              <input className={styles.boxunchecked2} type="radio"   name="radio" value={recordid4} checked={selectedRadio === recordid4} onChange={handleCheckboxChange} />
+              <input className={styles.boxunchecked2} type="radio"   name="radio" value={`${recordid4}_${subjectname4}`} checked={selectedRadio === recordid4} onChange={handleCheckboxChange} />
             </div> )}
             {(jValue+4 < datalength) && (
             <div className={styles.div12}>
@@ -327,7 +328,7 @@ const SeleccionEstudiante1 = () => {
                 <div className={styles.div13}>{subjectlab5? 'Si':'No'}</div>
                 <div className={styles.div14}>{subjectcredit5}</div>
               </div>
-              <input className={styles.boxunchecked5} type="radio"   name="radio" value={recordid5} checked={selectedRadio === recordid5} onChange={handleCheckboxChange}/>
+              <input className={styles.boxunchecked5} type="radio"   name="radio" value={`${recordid5}_${subjectname5}`} checked={selectedRadio === recordid5} onChange={handleCheckboxChange}/>
             </div>)}
             {(jValue+5 < datalength) && (
             <div className={styles.div15}>
@@ -337,7 +338,7 @@ const SeleccionEstudiante1 = () => {
               </div>
               <div className={styles.div16}>{subjectlab6? 'Si':'No'}</div>
               <div className={styles.div17}>{subjectcredit6}</div>
-              <input className={styles.boxunchecked3} type="radio"   name="radio" value={recordid6} checked={selectedRadio === recordid6} onChange={handleCheckboxChange}/>
+              <input className={styles.boxunchecked3} type="radio"   name="radio" value={`${recordid6}_${subjectname6}`} checked={selectedRadio === recordid6} onChange={handleCheckboxChange}/>
             </div> )}
             <b className={styles.retiroDeMateria}>Retiro de materia</b>
           </div>
